[sui-tool] remove sui-data-ingestion-core dependency
