+++ crates/sui-tool/src/formal_snapshot_util.rs
@@ -6,7 +6,7 @@ use anyhow::anyhow;
 use futures::{StreamExt, TryStreamExt};
 use std::sync::Arc;
 use std::sync::atomic::{AtomicU64, Ordering};
-use sui_data_ingestion_core::{CheckpointReader, create_remote_store_client};
+use sui_storage::object_store::util::{build_object_store, fetch_checkpoint};
 use sui_types::messages_checkpoint::{CheckpointSequenceNumber, VerifiedCheckpoint};
 use sui_types::storage::WriteStore;
 
@@ -20,14 +20,14 @@ pub(crate) async fn read_summaries_for_list_no_verify<S>(
 where
     S: WriteStore + Clone,
 {
-    let client = create_remote_store_client(ingestion_url, vec![], 60)?;
+    let client = build_object_store(&ingestion_url, vec![]);
     futures::stream::iter(checkpoints)
-        .map(|sq| CheckpointReader::fetch_from_object_store(&client, sq))
+        .map(|sq| fetch_checkpoint(&client, sq))
         .buffer_unordered(concurrency)
         .try_for_each(|checkpoint| {
             let result = store
                 .insert_checkpoint(&VerifiedCheckpoint::new_unchecked(
-                    checkpoint.0.checkpoint_summary.clone(),
+                    checkpoint.summary.clone(),
                 ))
                 .map_err(|e| anyhow!("Failed to insert checkpoint: {e}"));
             checkpoint_counter.fetch_add(1, Ordering::Relaxed);
